pyproject.toml
[project]
name = "llmops-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "awscli>=1.33.44",
    "boto3==1.34.37",
    "bs4>=0.0.2",
    "chromadb>=1.4.0",
    "fastapi>=0.128.0",
    "langchain>=1.2.3",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.7",
    "langchain-groq>=1.1.1",
    "langchainhub>=0.1.21",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.2.0",
    "tiktoken>=0.12.0",
    "uvicorn>=0.40.0",
]
